\name{bits.per.integer}
\alias{bits.per.integer}

\title{ Internal Size of an integer }
\description{
  bits.per.integer tells the number of bits that an integer occupies.
}
\usage{
bits.per.integer()
}

\details{
  Assigned to be 32 for the C programs
}

\author{ Knut M. Wittkowski \email{kmw@rockefeller.edu} }

\examples{
bits.per.integer()

## The function is currently defined as
function() 32
}

\keyword{ utilities }
