\name{muStat-package}
\alias{muStat-package}
\alias{muStat}
\docType{package}

\title{
Prentice (Friedman/Wilcoxon/Kruskal) Rank Sum Test,
Stratified McNemar (TDT, Sign) Test and u/mu-Scores for
Multivariate Data
}

\description{
  Performs a generalized Friedman rank sum test with replicated blocked
  data or, as special cases, a Kruskal-Wallis rank sum test on data 
  following a one-way layout, or a Wilcoxon rank sum test following 
  a one-way layout with only two groups, or a stratified McNemar/sign
  test using a notion adopted to genetic association studies (TDT).
  Generates u/mu-scores of univariate and multivariate data to be used 
  with these tests.
}

\section{Author}{
  Knut M. Wittkowski \email{kmw@rockefeller.edu}, 
  Tingting Song \email{ttsong@gmail.com}

Maintainer: Tingting Song \email{ttsong@gmail.com}
}

\section{Details}{
  \tabular{ll}{
    Package: \tab muStat\cr
    Type: \tab Package\cr
    Version: \tab 1.5.0\cr
    Date: \tab 2008-02-07\cr
    URL: \tab \url{http://mustat.rockefeller.edu} \cr
    License: \tab GPL (>= 2)
  }
}

\section{Copyright}{  
  \tabular{l}{\bold{Software Transfer Agreement} \cr}
  (adapted from the NIH Uniform Biological Materials 
    Transfer Agreement, 64 FR 72090)
  
  \

  In response to the RECIPIENT's request for the software package 
  \bold{muStat} \dQuote{SOFTWARE} the PROVIDER asks that the RECIPIENT 
  agree to the following:

  \enumerate{
    \item The SOFTWARE is the property of the PROVIDER and is made 
          available as a service to the research community.
    
    \item THE SOFTWARE WILL NOT BE USED FOR TREATING OR DIAGNOSING 
          HUMAN SUBJECTS.
  
    \item The SOFTWARE will be used for teaching or not-for-profit 
          research purposes only.  
  
    \item The SOFTWARE will not be further distributed to others 
          without the PROVIDER's written consent, including any 
          portion of the code incorporated into another work. 
          The RECIPIENT shall refer any request for the SOFTWARE to 
          the PROVIDER. PROVIDER agrees to make the SOFTWARE available, 
          under a separate Agreement to other scientists for teaching 
          or not-for-profit research purposes only.
    
    \item The RECIPIENT agrees to acknowledge the source of the 
          SOFTWARE in any publications reporting use of it. 
  
    \item THE PROVIDER MAKES NO REPRESENTATIONS AND EXTENDS NO 
          WARRANTIES OF ANY KIND, EITHER EXPRESSED OR IMPLIED. 
          THERE ARE NO EXPRESS OR IMPLIED WARRANTIES OF 
          MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE, 
          OR THAT THE USE OF THE SOFTWARE WILL NOT INFRINGE ANY 
          PATENT, COPYRIGHT, TRADEMARK, OR OTHER PROPRIETARY RIGHTS. 
          Unless prohibited by law, RECIPIENT assumes all liability 
          for claims for damages against it which may arise from the 
          use of the SOFTWARE.  
  
    \item By registering at \url{http://muStat.rockefeller.edu} 
          (\dQuote{PROVIDER SCIENTIST WEB SITE}) the RECIPIENT obtains the right 
          to receive revisions of the SOFTWARE and related information. 
  
    \item By downloading this SOFTWARE, the RECIPIENT agrees to 
          the terms of this AGREEMENT. 
  }
  

  \tabular{ll}{
  \bold{Provider Information:}\tab \cr
  Provider Scientist: \tab 
  \bold{Knut M. Wittkowski, PhD, DSc} \email{kmw@rockefeller.edu} \cr
  Provider: \tab 
  \bold{The Rockefeller University} \cr
  Address: \tab
  \bold{1230 York Ave, New York, NY 10021, U.S.A.} \cr
  Name of Authorized Official: \tab
  \bold{Kathleen Denis} \email{denisk@rockefeller.edu} \cr
  Title of Authorizing Official: \tab 
  \bold{Associate Vice President, Technology Transfer} \cr
  }
}

\keyword{ package }
